chore(parser): move getPatter to openapi
