+++ src/openApi/common/parser/__tests__/getPattern.spec.ts
