+++ src/openApi/common/parser/getPattern.ts
