+++ src/openApi/v2/parser/getModelProperties.ts
@@ -1,6 +1,6 @@
 import type { Model } from '../../../types/client';
 import { escapeName } from '../../../utils/escapeName';
-import { getPattern } from '../../../utils/getPattern';
+import { getPattern } from '../../common/parser/getPattern';
 import { getType } from '../../common/parser/type';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiSchema } from '../interfaces/OpenApiSchema';
