+++ src/openApi/v2/parser/getOperationParameter.ts
@@ -1,6 +1,6 @@
 import type { OperationParameter } from '../../../types/client';
 import { getEnums } from '../../../utils/getEnums';
-import { getPattern } from '../../../utils/getPattern';
+import { getPattern } from '../../common/parser/getPattern';
 import { getRef } from '../../common/parser/getRef';
 import { getOperationParameterName } from '../../common/parser/operation';
 import { getType } from '../../common/parser/type';
