+++ src/openApi/v3/parser/getOperationRequestBody.ts
@@ -1,5 +1,5 @@
 import type { OperationParameter } from '../../../types/client';
-import { getPattern } from '../../../utils/getPattern';
+import { getPattern } from '../../common/parser/getPattern';
 import { getType } from '../../common/parser/type';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiRequestBody } from '../interfaces/OpenApiRequestBody';
